playerの攻撃モーション等のアタッチ
