+++ Assets/Scenes/game_scene_each/NY'_scence/tagame90_transparent_0.prefab
@@ -12,7 +12,6 @@ GameObject:
   - component: {fileID: 2410179526740472512}
   - component: {fileID: 8533724371487852540}
   - component: {fileID: 6258827607541058357}
-  - component: {fileID: -2853051289036119998}
   m_Layer: 0
   m_Name: tagame90_transparent_0
   m_TagString: Untagged
@@ -153,16 +152,3 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 0
---- !u!114 &-2853051289036119998
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6564305981207148596}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8a78af05308f94b4e9779c9fa708b0de, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  explosionPrefab: {fileID: 0}
