--- Assets/Scenes/game_scene_each/kou's_scene/Ken_Skill.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 79b3473fcf0854e12a9c9f24ac6df0e9
