--- Assets/Scenes/game_scene_each/kou's_scene/Mito_Skill.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bee3286a32fb7441698d804d812c08d3
