--- Assets/Scenes/game_scene_each/kou's_scene/Sora_Skill.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3acabcd1290de4ae4b60e5b59503231e
