--- Assets/Scenes/game_scene_each/kou's_scene/Uruha_Skill.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d7aca49f794cc4d8696e8c202eed5c7b
