+++ Assets/Scripts/Player/Mito_Skill.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d744aa9604d8047adbc3badd077c2d90
