+++ Assets/Scripts/Player/Uruha_Skill.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f8265b34c7b4248d3b3a45c0efbcfcff
