+++ ProjectSettings/TagManager.asset
@@ -8,6 +8,7 @@ TagManager:
   - Enemy
   - wall
   - Goal
+  - Bullet
   layers:
   - Default
   - TransparentFX
